FF Upgrade support
The whole upgrade process is based on ISSU.
The main task of ff upgrade is to stop old
contrail control and those computes that sill have
old agent (it mught be nodes without workloads).
Because at the step of ff upgrade the ISSU sync is already
run, compute nodes that have workload are upgraded,
but they containers are tempopary because they run via docker_compose
At the latest step of ff upgrade these temporary agent containers
are to be replaced with new ones that are to be deloyed
by tripleo.

Change-Id: Ib97efce11e6314c0aeac223f4a55fc2c80bcadd1
Partial-Bug: #1801956