Enable installing with different container tags
Changing docker heat templates so that they can take advantage of
environments/contrail/contrail-docker-registry.yaml, where images
are given in format <registry>:<port>/<namespace>/<image_name>:<tag>

Normalisation is done by a separate resource
tools/contrail/contrail-image-normalize.yaml, that needs to be
included in each docker resource. Docker resources should take
its output as their images.

During normalisation we are checking if image contains registry
and tag. If some of these is missing, it is addedd to the image
form the global parameters ContrailImageTag and ContrailRegistry.

If image is empty, no normalisation is done and we just return
empty image.

Change-Id: I7dbd43de58524c3adc0d16d701eac589027a1e56
Closes-Bug: #1804213